refactor: DomainService 를 역할에 맞게 쪼개고 이름을 변경
