+++ src/main/java/kitchenpos/eatinorders/tobe/application/OrderService.java
@@ -1,6 +1,12 @@
 package kitchenpos.eatinorders.tobe.application;
 
-import kitchenpos.eatinorders.tobe.domain.*;
+import kitchenpos.eatinorders.tobe.domain.Order;
+import kitchenpos.eatinorders.tobe.domain.OrderLineItem;
+import kitchenpos.eatinorders.tobe.domain.OrderQuantity;
+import kitchenpos.eatinorders.tobe.domain.OrderRepository;
+import kitchenpos.eatinorders.tobe.domain.ordertable.OrderTableManager;
+import kitchenpos.eatinorders.tobe.domain.service.OrderDeliverService;
+import kitchenpos.eatinorders.tobe.domain.service.OrderValidateService;
 import kitchenpos.eatinorders.tobe.dto.CreateOrderRequest;
 import kitchenpos.eatinorders.tobe.dto.OrderCompletedResponse;
 import kitchenpos.eatinorders.tobe.dto.OrderLineItemRequest;
@@ -16,11 +22,15 @@ import java.util.stream.Collectors;
 @Service("TobeOrderService")
 public class OrderService {
     private final OrderRepository orderRepository;
-    private final OrderDomainService orderDomainService;
+    private final OrderValidateService orderValidateService;
+    private final OrderDeliverService orderDeliverService;
+    private final OrderTableManager orderTableManager;
 
-    public OrderService(final OrderRepository orderRepository, final OrderDomainService orderDomainService) {
+    public OrderService(final OrderRepository orderRepository, final OrderValidateService orderValidateService, final OrderDeliverService orderDeliverService, final OrderTableManager orderTableManager) {
         this.orderRepository = orderRepository;
-        this.orderDomainService = orderDomainService;
+        this.orderValidateService = orderValidateService;
+        this.orderDeliverService = orderDeliverService;
+        this.orderTableManager = orderTableManager;
     }
 
     @Transactional
@@ -35,7 +45,7 @@ public class OrderService {
                 request.getDeliveryAddress(),
                 request.getOrderTableId()
         );
-        orderDomainService.validateOrder(order);
+        orderValidateService.validateOrder(order);
         orderRepository.save(order);
         return createOrderResponse(order);
     }
@@ -44,7 +54,7 @@ public class OrderService {
     public OrderResponse accept(final UUID orderId) {
         final Order order = orderRepository.findById(orderId)
                 .orElseThrow(NoSuchElementException::new);
-        orderDomainService.acceptOrder(order);
+        order.accept(orderDeliverService);
         return createOrderResponse(order);
     }
 
@@ -52,7 +62,7 @@ public class OrderService {
     public OrderResponse serve(final UUID orderId) {
         final Order order = orderRepository.findById(orderId)
                 .orElseThrow(NoSuchElementException::new);
-        orderDomainService.serveOrder(order);
+        order.serve();
         return createOrderResponse(order);
     }
 
@@ -60,7 +70,7 @@ public class OrderService {
     public OrderResponse startDelivery(final UUID orderId) {
         final Order order = orderRepository.findById(orderId)
                 .orElseThrow(NoSuchElementException::new);
-        orderDomainService.startDelivery(order);
+        order.startDelivery();
         return createOrderResponse(order);
     }
 
@@ -68,7 +78,7 @@ public class OrderService {
     public OrderResponse completeDelivery(final UUID orderId) {
         final Order order = orderRepository.findById(orderId)
                 .orElseThrow(NoSuchElementException::new);
-        orderDomainService.completeDelivery(order);
+        order.completeDelivery();
         return createOrderResponse(order);
     }
 
@@ -76,7 +86,7 @@ public class OrderService {
     public OrderResponse complete(final UUID orderId) {
         final Order order = orderRepository.findById(orderId)
                 .orElseThrow(NoSuchElementException::new);
-        orderDomainService.completeOrder(order);
+        order.complete(orderTableManager);
         return createOrderResponse(order);
     }
 
@@ -98,7 +108,7 @@ public class OrderService {
                 order.getOrderDateTime(),
                 order.getOrderLineItems(),
                 order.getDeliveryAddress(),
-                orderDomainService.getOrderTable(order),
+                orderTableManager.getOrderTable(order.getOrderTableId()),
                 order.getOrderTableId()
         );
     }
